Added styles for header, main, and footer.
For the header, I resized the logo image, so that it is proportionate and not too big or small.

The header navigation links now float to the right, and hovering over them will have a gray padding to give visual feedback that they work.

The header and main navigation links are separate, the header links do not have an underline to make it look proffesional. The main navigation links have an underline and also have a dark padding for visual feedback.

Imported 'Inter' font from Google.
I tweaked the font a little bit to make it thin, it fits the aesthetic of the website.

I used 'safety orange' for the sign up links in the main section to make them pop and easy to see.

Paddings have rounded corners.